Avoid having move meta info in Move
The move meta information like turnIndex, playerIndex, and movesRemaining belongs in the state, not the move itself.
For now the move instance is removed from the state because it was only there to hold meta info.
It will be replaced shortly with the move to take to transform to the next state.
z -> conqueredRegions